// dqops/src/main/frontend/src/redux/types/definition.types.ts
export enum DEFINITION_ACTION {
  GET_SENSOR_FOLDER_TREE = 'DEFINITION_ACTION/GET_SENSOR_FOLDER_TREE',
  GET_SENSOR_FOLDER_TREE_SUCCESS = 'DEFINITION_ACTION/GET_SENSOR_FOLDER_TREE_SUCCESS',
  GET_SENSOR_FOLDER_TREE_ERROR = 'DEFINITION_ACTION/GET_SENSOR_FOLDER_TREE_ERROR',

  UPDATE_SENSOR_FOLDER_TREE = 'DEFINITION_ACTION/UPDATE_SENSOR_FOLDER_TREE',
  TOGGLE_SENSOR_FOLDER = 'DEFINITION_ACTION/TOGGLE_SENSOR_FOLDER',
  OPEN_SENSOR_FOLDER = 'DEFINITION_ACTION/OPEN_SENSOR_FOLDER',
  ADD_FIRST_LEVEL_TAB = 'DEFINITION_ACTION/ADD_FIRST_LEVEL_TAB',
  SET_ACTIVE_FIRST_LEVEL_TAB = 'DEFINITION_ACTION/SET_ACTIVE_FIRST_LEVEL_TAB',
  CLOSE_FIRST_LEVEL_TAB = 'DEFINITION_ACTION/CLOSE_FIRST_LEVEL_TAB',
  GET_SENSOR_DETAIL = 'DEFINITION_ACTION/GET_SENSOR_DETAIL',
  GET_SENSOR_DETAIL_SUCCESS = 'DEFINITION_ACTION/GET_SENSOR_DETAIL_SUCCESS',
  GET_SENSOR_DETAIL_FAILED = 'DEFINITION_ACTION/GET_SENSOR_DETAIL_FAILED',
  GET_RULE_DETAIL = 'DEFINITION_ACTION/GET_RULE_DETAIL',
  GET_RULE_DETAIL_SUCCESS = 'DEFINITION_ACTION/GET_RULE_DETAIL_SUCCESS',
  GET_RULE_DETAIL_FAILED = 'DEFINITION_ACTION/GET_RULE_DETAIL_FAILED',

  SET_UPDATED_RULE = 'DEFINITION_ACTION/SET_UPDATED_RULE',
  SET_UPDATED_SENSOR = 'DEFINITION_ACTION/SET_UPDATED_SENSOR',

  UPDATE_SENSOR_DETAIL = 'DEFINITION_ACTION/UPDATE_SENSOR_DETAIL',
  UPDATE_SENSOR_DETAIL_SUCCESS = 'DEFINITION_ACTION/UPDATE_SENSOR_DETAIL_SUCCESS',
  UPDATE_SENSOR_DETAIL_FAILED = 'DEFINITION_ACTION/UPDATE_SENSOR_DETAIL_FAILED',

  UPDATE_RULE_DETAIL = 'DEFINITION_ACTION/UPDATE_RULE_DETAIL',
  UPDATE_RULE_DETAIL_SUCCESS = 'DEFINITION_ACTION/UPDATE_RULE_DETAIL_SUCCESS',
  UPDATE_RULE_DETAIL_FAILED = 'DEFINITION_ACTION/UPDATE_RULE_DETAIL_FAILED',

  CREATE_RULE_DETAIL = 'DEFINITION_ACTION/CREATE_RULE_DETAIL',
  CREATE_RULE_DETAIL_SUCCESS = 'DEFINITION_ACTION/CREATE_RULE_DETAIL_SUCCESS',
  CREATE_RULE_DETAIL_FAILED = 'DEFINITION_ACTION/CREATE_RULE_DETAIL_FAILED',

  CREATE_SENSOR_DETAIL = 'DEFINITION_ACTION/CREATE_SENSOR_DETAIL',
  CREATE_SENSOR_DETAIL_SUCCESS = 'DEFINITION_ACTION/CREATE_SENSOR_DETAIL_SUCCESS',
  CREATE_SENSOR_DETAIL_FAILED = 'DEFINITION_ACTION/CREATE_SENSOR_DETAIL_FAILED',
  GET_DATA_QUALITY_CHECKS_FOLDER_TREE = 'DEFINITION_ACTION/GET_DATA_QUALITY_CHECKS_FOLDER_TREE',
  GET_DATA_QUALITY_CHECKS_FOLDER_TREE_SUCCESS = 'DEFINITION_ACTION/GET_DATA_QUALITY_CHECKS_FOLDER_TREE_SUCCESS',
  GET_DATA_QUALITY_CHECKS_FOLDER_TREE_ERROR = 'DEFINITION_ACTION/GET_DATA_QUALITY_CHECKS_FOLDER_TREE_ERROR',
  TOGGLE_DATA_QUALITY_CHECKS_FOLDER = 'DEFINITION_ACTION/TOGGLE_DATA_QUALITY_CHECKS_FOLDER',
  UPDATE_DATA_QUALITY_CHECKS_FOLDER_TREE = 'DEFINITION_ACTION/UPDATE_DATA_QUALITY_CHECKS_FOLDER_TREE',
  OPEN_DATA_QUALITY_CHECKS_FOLDER = 'DEFINITION_ACTION/OPEN_DATA_QUALITY_CHECKS_FOLDER',

  CREATE_CHECK_DETAIL = 'DEFINITION_ACTION/CREATE_CHECK_DETAIL',
  UPDATE_CHECK_DETAIL = 'DEFINITION_ACTION/UPDATE_CHECK_DETAIL',
  DELETE_CHECK_DETAIL = 'DEFINITION_ACTION/DELETE_CHECK_DETAIL',
  GET_CHECK_DETAIL = 'DEFINITION_ACTION/GET_CHECK_DETAIL',
  GET_CHECK_DETAIL_SUCCESS = 'DEFINITION_ACTION/GET_CHECK_DETAIL_SUCCESS',

  GET_RULE_FOLDER_TREE = 'DEFINITION_ACTION/GET_RULE_FOLDER_TREE',
  GET_RULE_FOLDER_TREE_SUCCESS = 'DEFINITION_ACTION/GET_RULE_FOLDER_TREE_SUCCESS',
  GET_RULE_FOLDER_TREE_ERROR = 'DEFINITION_ACTION/GET_RULE_FOLDER_TREE_ERROR',
  TOGGLE_RULE_FOLDER = 'DEFINITION_ACTION/TOGGLE_RULE_FOLDER',
  UPDATE_RULE_FOLDER_TREE = 'DEFINITION_ACTION/UPDATE_RULE_FOLDER_TREE',
  OPEN_RULE_FOLDER = 'DEFINITION_ACTION/OPEN_RULE_FOLDER',

  TOGGLE_FIRST_LEVEL_FOLDER = 'DEFINITION/TOGGLE_FIRST_LEVEL_FOLDER',

  REFRESH_CHECKS_TREE_INDICATOR = 'DEFINITION_ACTION/REFRESH_CHECKS_TREE_INDICATOR',
  REFRESH_SENSORS_TREE_INDICATOR = 'DEFINITION_ACTION/REFRESH_SENSORS_TREE_INDICATOR',
  REFRESH_RULES_TREE_INDICATOR = 'DEFINITION_ACTION/REFRESH_RULES_TREE_INDICATOR',


}